+++ turretchassisdef.go
@@ -19,26 +19,26 @@ type TurretChassisDef struct {
 	} `json:"Hardpoints"`
 	InventorySlots     int64 `json:"InventorySlots"`
 	LOSSourcePositions []struct {
-		X int64 `json:"x"`
-		Y int64 `json:"y"`
-		Z int64 `json:"z"`
+		X float64 `json:"x"`
+		Y float64 `json:"y"`
+		Z float64 `json:"z"`
 	} `json:"LOSSourcePositions"`
 	LOSTargetPositions []struct {
-		X int64 `json:"x"`
-		Y int64 `json:"y"`
-		Z int64 `json:"z"`
+		X float64 `json:"x"`
+		Y float64 `json:"y"`
+		Z float64 `json:"z"`
 	} `json:"LOSTargetPositions"`
-	MaxArmor                  int64  `json:"MaxArmor"`
-	MaxInternalStructure      int64  `json:"MaxInternalStructure"`
-	MovementCapDefID          string `json:"MovementCapDefID"`
-	PathingCapDefID           string `json:"PathingCapDefID"`
-	PrefabBase                string `json:"PrefabBase"`
-	PrefabIdentifier          string `json:"PrefabIdentifier"`
-	Radius                    int64  `json:"Radius"`
-	SensorRangeMultiplier     int64  `json:"SensorRangeMultiplier"`
-	Signature                 int64  `json:"Signature"`
-	SpotterDistanceMultiplier int64  `json:"SpotterDistanceMultiplier"`
-	Tonnage                   int64  `json:"Tonnage"`
-	VisibilityMultiplier      int64  `json:"VisibilityMultiplier"`
-	WeightClass               string `json:"weightClass"`
+	MaxArmor                  int64   `json:"MaxArmor"`
+	MaxInternalStructure      int64   `json:"MaxInternalStructure"`
+	MovementCapDefID          string  `json:"MovementCapDefID"`
+	PathingCapDefID           string  `json:"PathingCapDefID"`
+	PrefabBase                string  `json:"PrefabBase"`
+	PrefabIdentifier          string  `json:"PrefabIdentifier"`
+	Radius                    int64   `json:"Radius"`
+	SensorRangeMultiplier     float64 `json:"SensorRangeMultiplier"`
+	Signature                 float64 `json:"Signature"`
+	SpotterDistanceMultiplier float64 `json:"SpotterDistanceMultiplier"`
+	Tonnage                   int64   `json:"Tonnage"`
+	VisibilityMultiplier      float64 `json:"VisibilityMultiplier"`
+	WeightClass               string  `json:"weightClass"`
 }
